+++ bindgen/src/gen_cs/custom.rs
@@ -3,16 +3,17 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use super::CodeType;
-use uniffi_bindgen::{backend::Literal, ComponentInterface};
+use uniffi_bindgen::{interface::Literal, ComponentInterface};
 
 #[derive(Debug)]
 pub struct CustomCodeType {
     name: String,
+    builtin: Box<dyn CodeType>,
 }
 
 impl CustomCodeType {
-    pub fn new(name: String) -> Self {
-        CustomCodeType { name }
+    pub fn new(name: String, builtin: Box<dyn CodeType>) -> Self {
+        CustomCodeType { name, builtin }
     }
 }
 
@@ -27,5 +28,9 @@ impl CodeType for CustomCodeType {
 
     fn literal(&self, _literal: &Literal, _ci: &ComponentInterface) -> String {
         unreachable!("Can't have a literal of a custom type");
+    }
+
+    fn default_value(&self, ci: &ComponentInterface) -> String {
+        self.builtin.default_value(ci)
     }
 }